splitting some code for cleanup
